examples/python/requirements.txt
mcp
smolagents[litellm]
rossum-api @ git+https://github.com/rossumai/rossum-sdk#egg=rossum-api
rossum-mcp @ git+https://github.com/stancld/rossum-mcp#egg=rossum-mcp
pyyaml
plotly>=5.18.0
kaleido>=0.2.1
matplotlib>=3.8.0
seaborn>=0.13.0
numpy>=1.26.0
streamlit>=1.50.0
